Font adjustments
Changed from sp to dp given how sensitive layout is to font size.  Made
fonts smaller across the board since there doesn't seem to be anything
beyond dp to make fonts smaller on smaller screens.  GUI now looks good
on 7" Kindle Fire.  It's not quite as nice as it used to be on 10"
Galaxy Tab, but it's still very usable and if you didn't know better,
you wouldn't know anything is off.  Fixes #10.